fix(rpc): rename debug test file and remove unused helpers
Rename debug_trace_tests.rs to debug_preimage_tests.rs to match its
actual content (only tests debug_preimage). Remove unused scaffold
helpers (build_block, create_transfer_tx, build_and_execute_block,
setup_single_transfer_block, TestEnv) that produced dead-code warnings.
