+++ test/tests/rpc/mod.rs
@@ -1,6 +1,7 @@
 mod authrpc_batch_tests;
 mod block_access_list_tests;
 mod client_version_tests;
+mod debug_preimage_tests;
 mod debug_trace_tests;
 mod fork_choice_tests;
 mod http_batch_tests;
